+++ CHANGELOG.md
@@ -23,18 +23,22 @@ Here is a template for new release sections
 
 ### Added
 - Instruction to install graphviz on windows in `docs/troubleshooting.rst` (#572)
-
+- Benchmark test `test_benchmark_feature_parameters_as_timeseries` to ensure that parameters can always also be defined as a timeseries. Applied to `efficiency` of an energyConversion asset and `electricity_price` of an energyProduction asset (#542)
+- Input files for benchmark tests `test_benchmark_feature_input_flows_as_list` (`Feature_input_flows_as_list`) and `test_benchmark_feature_output_flows_as_list` (`Feature_output_flows_as_list`), but not the benchmark assertions (#542)
 
 ### Changed
 - Modify `setup.py` to upload the code as package on pypi.org (#570)
 - Improve message when the `tests/test_input_folder_parameters.py` fails (#578)
 - Modify PR template to precise to add assert message and link to example docstring 
 - Update CONTRIBUTING to add a "Write test for your code" section before the "Run tests locally" one (#579)
+- Changed `E2.calculate_dispatch_expenditures()` so that it can process parameters defined as lists (#542)
 
 ### Removed
 -
+
 ### Fixed
 - Update the release protocol in `CONTRIBUTING.md` file (#576)
+- Fix reading timeseries for parameters in `C0` (#542)
 
 ## [0.4.1] - 2020-09-21